Struct accesskit_macos::NSArray

pub struct NSArray<T, O = Shared>where
    T: Message,
    O: Ownership,{ /* private fields */ }

An immutable ordered collection of objects.

This is the Objective-C equivalent of a “boxed slice” (Box<[T]>), so effectively a Vec<T> where you can’t change the number of elements.

The type of the contained objects is described by the generic parameter T, and the ownership of the objects is described with the generic parameter O.

Ownership

While NSArray itself is immutable, i.e. the number of objects it contains can’t change, it is still possible to modify the contained objects themselves, if you know you’re the sole owner of them - quite similar to how you can modify elements in Box<[T]>.

To mutate the contained objects the ownership must be O = Owned. A summary of what the different “types” of arrays allow you to do can be found below. Array refers to either NSArray or NSMutableArray.

Id<NSMutableArray<T, Owned>, Owned>: Allows you to mutate the objects, and the array itself.

Id<NSMutableArray<T, Shared>, Owned>: Allows you to mutate the array itself, but not it’s contents.

Id<NSArray<T, Owned>, Owned>: Allows you to mutate the objects, but not the array itself.

Id<NSArray<T, Shared>, Owned>: Effectively the same as the below.

Id<Array<T, Shared>, Shared>: Allows you to copy the array, but does not allow you to modify it in any way.

Id<Array<T, Owned>, Shared>: Pretty useless compared to the others, avoid this.

See Apple’s documentation.

Implementations

impl<T, O> NSArray<T, O>where
T: Message,
O: Ownership,

Generic creation methods.

pub fn new() -> Id<NSArray<T, O>, Shared>

Get an empty array.

pub fn from_vec(vec: Vec<Id<T, O>, Global>) -> Id<NSArray<T, O>, O>

impl<T> NSArray<T, Shared>where
T: Message,

Creation methods that produce shared arrays.

pub fn from_slice(slice: &[Id<T, Shared>]) -> Id<NSArray<T, Shared>, Shared>

impl<T, O> NSArray<T, O>where
T: Message,
O: Ownership,

Generic accessor methods.

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn get(&self, index: usize) -> Option<&T>

pub fn first(&self) -> Option<&T>

pub fn last(&self) -> Option<&T>

pub fn iter(&self) -> NSEnumerator<'_, T>

pub fn objects_in_range(&self, range: Range<usize>) -> Vec<&T, Global>

pub fn to_vec(&self) -> Vec<&T, Global>

pub fn into_vec(array: Id<NSArray<T, O>, Owned>) -> Vec<Id<T, O>, Global>

impl<T> NSArray<T, Shared>where
T: Message,

Accessor methods that work on shared arrays.

pub fn get_retained(&self, index: usize) -> Id<T, Shared>

pub fn to_shared_vec(&self) -> Vec<Id<T, Shared>, Global>

impl<T> NSArray<T, Owned>where
T: Message,

Accessor methods that work on owned arrays.

pub fn get_mut(&mut self, index: usize) -> Option<&mut T>

pub fn first_mut(&mut self) -> Option<&mut T>

pub fn last_mut(&mut self) -> Option<&mut T>